refactor(reqresp): split monolith into messages, codec, server, client
